Add automated API tests using schemathesis
This adds support for testing the documentation against the API.
The data sent in the requests are based on the examples we define in openapi.yml.

Schemathesis checks the response code, content type and the schema of the json response.
The test fails if none of it matches what is in the OpenAPI spec.

The request and response body is logged in build/cassette.yml,
and the output is set to verbose mode so all the data can be seen on stdout.