ci: add client libraries generation rules
This adds the commands to generate client libraries from the openapi spec
to the Makefile. Supported languages are (python, c#, go and typescript).

For the Actions pipelines to work, `GIT_EMAIL` and `GIT_USERNAME` should be
configured as GitHub secret tokens as they are needed for the `git commit` commands to work